handle points better.
Simply updating the points data runs into problems if processing a
roi removes points, since they would simply be added back in by the
merge.
New behavior is to replace points in the smaller roi.
If copy_from_self is True, and self is the larger roi, self remains unchanged.
Otherwise all points in self in the smaller roi are deleted, and then
all nodes in other are filled in.